+++ packages/crdt_lf_flutter/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.3.0](https://github.com/MattiaPispisa/crdt/tree/crdt_lf_flutter-v0.3.0/packages/crdt_lf_flutter)
+
+**Date:** 2026-07-30
+
+[compare to previous release](https://github.com/MattiaPispisa/crdt/compare/crdt_lf_flutter-v0.2.1...crdt_lf_flutter-v0.3.0)
+
+### Added
+
+- `CrdtTextCursorsOverlay` carets now moves to their new position instead of jumping there. [102](https://github.com/MattiaPispisa/crdt/issues/102)
+
 ## [0.2.1](https://github.com/MattiaPispisa/crdt/tree/crdt_lf_flutter-v0.2.1/packages/crdt_lf_flutter)
 
 **Date:** 2026-07-21